refactor: move coins to CashHolder
